+++ vistas/login.html
@@ -92,7 +92,7 @@
 <div id="login-wrapper">
 
     <div class="u-form u-radius-20 u-white u-form-1" id="loginForm">
-        <form  action="/clientes/login" method="POST" class="u-clearfix u-form-spacing-15 u-form-vertical u-inner-form" style="padding: 23px;">
+        <form  action="/login/autenticacion" method="POST" class="u-clearfix u-form-spacing-15 u-form-vertical u-inner-form" style="padding: 23px;">
             <h4 class="u-align-center u-form-group u-form-text u-text u-text-2">Iniciar sesión </h4>
             <div class="u-form-group u-form-name">
                 <label class="u-custom-font u-font-roboto u-label">Correo</label>
